Refactor secure routing into own class
